Change requirejs_include_tag policy to one-per-page.
The original concept behind requirejs_include_tag was for it to act as
an replacement for javascript_include_tag.  Since a significant part of
its role is to emit the require.js configuration and the tag that
includes require.js itself, it must come first on a page.  However,
there's no definition as to which helper instance is "topmost".

This commit puts a stake in the ground: only one requirejs_include_tag
can exist on the page.  It's the user's responsibility to get this order
correct via their layout structure.  We die hard with an exception if
there's any ambiguity on this point.