Fix requirejs_include_tag regression in production
When `paths` is specified in `config/requirejs.yml`,
requirejs_include_tag would incorrectly attempt to map digestified paths
to the RHS of those assets.  That was unecessary work on the gem's part,
since any modules intended to exist in the layered build must be
explicitly specified in the `modules` section of requirejs.yml.