Update requirejs_helper.rb
Fixes #182 - Per the discussion in https://github.com/codeforamerica/ohana-web-search/issues/552, the baseUrl appears to be overwritten for module paths unless there is a baseUrl specified before loading the require.js file. I don't think this is probably the best solution, but this gets the discussion going and does solve the problem for our setup.