Support custom Sprocket manifests
This change adds support for Rails 4 apps that have set their manifest to a
custom one.

In Rails 4, if `config.assets.manifest` is set, it must include the
manifest's filename. For example:

    config.assets.manifest = "public/assets/manifest-foo.json"

Note that while it's invalid for the manifest option to not include the filename
in Rails 4, we are more forgiving and also accept values that don't contain
the filename, in order to preserve backwards compatibility with previous
requirejs-rails versions and Rails 3 apps.