Fixes `config.assets.manifest` issue
== Problem

Setting the Rails config `config.assets.manifest` using a recent
sprockets version (>= 3.0) makes the use of Requirejs::Rails impossible.

Sprockets change the content of the configuration from being a directory
to a file. Requirejs::Rails was still expecting a directory.

== Solution

Get the dirname of the given value if this value is a path to a file.